refactor: Move tests directory
