+++ eslint.config.js
@@ -1,14 +1,6 @@
 // @ts-check
 
-// @ts-ignore Needed due to moduleResolution Node vs Bundler
+// @ts-expect-error
 import { tanstackConfig } from '@tanstack/config/eslint'
 
-export default [
-  ...tanstackConfig,
-  {
-    name: 'tanstack/local',
-    rules: {
-      'ts/prefer-for-of': 'off',
-    },
-  },
-]
+export default [...tanstackConfig]
