+++ package.json
@@ -38,7 +38,7 @@
   "devDependencies": {
     "@eslint-react/eslint-plugin": "^1.5.16",
     "@solidjs/testing-library": "^0.8.8",
-    "@tanstack/config": "^0.8.6",
+    "@tanstack/config": "^0.9.0",
     "@testing-library/jest-dom": "^6.4.6",
     "@testing-library/react": "^16.0.0",
     "@testing-library/user-event": "^14.5.2",
