+++ packages/angular-store/vite.config.ts
@@ -1,14 +1,15 @@
 import { defineConfig } from 'vitest/config'
+import packageJson from './package.json'
 
 export default defineConfig({
   test: {
-    name: 'angular-store',
+    name: packageJson.name,
     dir: './tests',
     watch: false,
     environment: 'jsdom',
-    setupFiles: ['tests/test-setup.ts'],
-    globals: true,
+    setupFiles: ['./tests/test-setup.ts'],
     coverage: { enabled: true, provider: 'istanbul', include: ['src/**/*'] },
     typecheck: { enabled: true },
+    globals: true,
   },
 })
