+++ packages/solid-store/tsconfig.legacy.json
@@ -7,6 +7,5 @@
       "@tanstack/store": ["../store/src"]
     }
   },
-  "include": ["src"],
-  "exclude": ["src/tests/**"]
+  "include": ["src"]
 }
