+++ packages/store/tsconfig.json
@@ -3,5 +3,5 @@
   "compilerOptions": {
     "moduleResolution": "Bundler"
   },
-  "include": ["src", "eslint.config.js", "vite.config.ts"]
+  "include": ["src", "tests", "eslint.config.js", "vite.config.ts"]
 }
