+++ packages/store/tsconfig.legacy.json
@@ -1,5 +1,4 @@
 {
   "extends": "../../tsconfig.json",
-  "include": ["src"],
-  "exclude": ["src/tests/**"]
+  "include": ["src"]
 }
