+++ packages/svelte-store/tsconfig.json
@@ -7,8 +7,11 @@
     "src/**/*.js",
     "src/**/*.ts",
     "src/**/*.svelte",
+    "tests/**/*.js",
+    "tests/**/*.ts",
+    "tests/**/*.svelte",
     "eslint.config.js",
-    "test-setup.ts",
+    "svelte.config.js",
     "vite.config.ts"
   ]
 }
