+++ packages/vue-store/package.json
@@ -32,12 +32,8 @@
   },
   "sideEffects": false,
   "scripts": {
-    "test:lib": "pnpm run test:3",
-    "test:2": "vue-demi-switch 2 vue2 && vitest",
-    "test:2.7": "vue-demi-switch 2.7 vue2.7 && vitest",
-    "test:3": "vue-demi-switch 3 && vitest",
     "clean": "rimraf ./dist && rimraf ./coverage",
-    "test:eslint": "eslint ./src",
+    "test:eslint": "eslint ./src ./tests",
     "test:types": "pnpm run \"/^test:types:ts[0-9]{2}$/\"",
     "test:types:ts49": "node ../../node_modules/typescript49/lib/tsc.js -p tsconfig.legacy.json",
     "test:types:ts50": "node ../../node_modules/typescript50/lib/tsc.js",
@@ -45,7 +41,11 @@
     "test:types:ts52": "node ../../node_modules/typescript52/lib/tsc.js",
     "test:types:ts53": "node ../../node_modules/typescript53/lib/tsc.js",
     "test:types:ts54": "tsc",
-    "fixme:test:lib": "pnpm run test:2 && pnpm run test:2.7 && pnpm run test:3",
+    "test:lib": "pnpm run test:lib:3",
+    "test:lib:2": "vue-demi-switch 2 vue2 && vitest",
+    "test:lib:2.7": "vue-demi-switch 2.7 vue2.7 && vitest",
+    "test:lib:3": "vue-demi-switch 3 && vitest",
+    "test:lib:fixme": "pnpm run test:lib:2 && pnpm run test:lib:2.7 && pnpm run test:lib:3",
     "test:lib:dev": "pnpm run test:lib --watch",
     "test:build": "publint --strict",
     "build": "vite build"
