+++ packages/vue-store/tsconfig.json
@@ -9,5 +9,5 @@
       "@tanstack/store": ["../store/src"]
     }
   },
-  "include": ["src", "eslint.config.js", "test-setup.ts", "vite.config.ts"]
+  "include": ["src", "tests", "eslint.config.js", "vite.config.ts"]
 }
