+++ packages/vue-store/tsconfig.legacy.json
@@ -8,6 +8,5 @@
       "@tanstack/store": ["../store/src"]
     }
   },
-  "include": ["src"],
-  "exclude": ["src/tests/**"]
+  "include": ["src"]
 }
